runtime:  static lock ranking for the runtime
I took some of the infrastructure from Austin's lock logging CR
https://go-review.googlesource.com/c/go/+/192704 (with deadlock detection from the
logs), and developed a setup to give static lock ranking for runtime locks.

Static lock ranking gives better runtime checking for possible deadlocks, since
you don't have to run into the deadlock -- you just have to run into a lock
acquisition sequence that violates the lock ranking. Along the way, I found a
deadlock involving the new timer code, which Ian fixed via
https://go-review.googlesource.com/c/go/+/207348

See the lockNames[] array in runtime/locklog.go to show the static lock ranking
that I ended up with, along with some comments. I think this is great
documentation of the current intended lock ordering when acquiring multiple locks
in the runtime.

I didn't yet try to check the ranking within a lock class with multiple locks that
can be acquired at the same time (i.e. check the ranking when multiple hchan locks
are acquired).

Currently, I am doing a lockInit() call to set the lock rank of most locks. Any
lock that is not otherwise initialized is assumes to be a leaf lock (a very high
rank lock), so that eliminates the need to do anything for a bunch of locks
(including a bunch of architecture-dependent locks). For two locks, root.lock and
notifyList.lock (only in the runtime/sema.go file), it is not as easy to do lock
initialization, so instead, I am passing the lock rank with the lock calls.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I154217ff307c47051f8dae9c2a03b53081acd83a