runtime:  static lock ranking for the runtime
I took some of the infrastructure from Austin's lock logging CR
https://go-review.googlesource.com/c/go/+/192704 (with deadlock detection from the
logs), and developed a setup to give static lock ranking for runtime locks.

Static lock ranking gives better runtime checking for possible deadlocks, since
you don't have to run into the deadlock -- you just have to run into a lock
acquisition sequence that violates the lock ranking. Along the way, I found a
deadlock involving the timers, that Ian is fixing via
https://go-review.googlesource.com/c/go/+/207348 (and is currently included in
this change so the ranking works).

See the lockNames[] array in locklog.go to show the static lock ranking that I
ended up with, along with some comments. I think this is great documentation of
the current intended lock ordering when acquiring multiple locks in the runtime.

I didn't yet try to check the ranking within a lock class with multiple locks that
can be acquired at the same time (i.e. check the ranking when multiple hchan locks
are acquired).

Also, we would need to discuss if/how to incorporate this change into the runtime
code permanently, since there is no lock init function to set the rank of a lock
(mutex). So, like Austin's change, I am instead passing the lock rank with every
lock call. I'm not sure if that is acceptable, or if we should try to make a mutex
init function.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I154217ff307c47051f8dae9c2a03b53081acd83a