Update patch set 1
Patch Set 1:

> 
> Why is the number of prefixes something the user would want to specify?
> (How can/should they decide where to set the prefix limit?)
> 
> 

The currently hard-coded maximum prefix limit of 5 is the recommended setting for affected processors (section 3.11 of the white paper).  However, the same page of the white paper (Page 11 footnote 2) notes that this setting may not be optimal for some un-affected processors. If you wanted to create one binary to run on affected and un-affected processors, you may chose to deviate from the recommended setting.

Patch-set: 1