[dev.link] cmd/link: initial support for linker-materialized external symbols
Create loader interfaces for constructing the payloads of external
symbols from scratch, as opposed to passing in a sym.Symbol object
containing the payload.

The general idea is that the loader client creates a new Sym using
loader.AddExtSym, and then can add relocations/data to that sym
using loader.{AddBytes,AddReloc} (and others).

This change preserves compatibility with the old way of doing things
(passing in sym.Symbol) via a new loader.InstallSym method. If a
client invokes this method for a specific Sym, then the loader keeps
track of this fact and uses the sym.Symbol as the backing store. So
for the moment we can have a "mix and match" situation.

Also included is a small unit test for the new interfaces -- not clear
whether this really needs to be kept around long term... it was mainly
useful during initial bringup.

Change-Id: If8ab15df7b64636e56b317155dfe6d7cdfe23b71