[dev.link] cmd/link/internal/loader: add more interfaces for updating symbol contents
DO NOT SUBMIT

Add in more interfaces/hooks for adding bits of data to a materialized
symbol. This is intended to provide a set of hooks that are mostly
compatible with existing sym.Symbol methods, but work with loader.Sym
global sym indices instead.

Not fully cooked, needs unit testing.

Change-Id: Ib4521eabe953dcde6235dd02425d0b3e9284a2dd