[dev.link] cmd/link: stop using syms to make new symbols
We need to push the wave front of sym.Lookup till later. As such, we
need to eliminate sym.Symbols from most entries into the loader. The
first step to this is to remove the dependecy of the loader on
sym.Newsym. From here we can:

1) Remove the dependencies of the host object loader on symbols.
2) Change the usage of *sym.Symbol to modify data to a different API.
   Right now, we're thinking of going through the loader, maybe
   something like:

     s := l.LookupOrCreate(foo, bar)
     l.SetType(s, STLSBSS)
     l.SetAttrib(s, ReadOnly)
     // etc.

From here, I'll add support for the new symbol modification API to host
objects loaders, linksetup, and dostrdata. These three places are what
stand between us and DWARF with the new loader API.

Change-Id: Iae60100539155c140eea24bc896b9b7f16bdb064