runtime: wake scavenger and update address on sweep done
This change modifies the semantics of waking the scavenger: rather than
wake on any update to pacing, wake when we know we will have work to do,
that is, when the sweeper is done. This provides a much closer match to
the current scavenger, which just runs over the whole address space
once. We don't want to wake it too early, because then it will go to
sleep and accumulate a backlog of work. If we wake it too late, then we
spend more time with unscavenged memory. The point at which the sweeper
is done is the ideal time to wake the scavenger, because we just
finished generating new memory. In applications that go idle, this means
the background scavenger will be triggered (if it's parked) even if
another GC doesn't happen, which isn't true today.

However, we're unable to wake the scavenger directly from within the
sweeper; waking the scavenger involves modifying timers and reading
goroutines, the latter of which may trigger an allocation today (and the
sweeper may run during allocation!). Instead, we do the following:

1. Set a flag which is checked by sysmon. sysmon will clear the flag and
   wake the scavenger.
2. Wake the scavenger unconditionally at sweep termination.

The idea behind this policy is that it gets us close enough to the state
above without having to deal with the complexity of waking the scavenger
in deep parts of the runtime. If the application goes idle and sweeping
finishes (so we don't reach sweep termination), then sysmon will wake
the scavenger. sysmon has a worst-case 20 ms delay in responding to this
signal, which is probably fine if the application is completely idle
anyway, but if the application is actively allocating, then the
proportional sweeper should help ensure that sweeping ends very close to
sweep termination, so sweep termination is a perfectly reasonable time
to wake up the scavenger.

Updates #35788.

Change-Id: I84289b37816a7d595d803c72a71b7f5c59d47e6b