[dev.link] cmd/link: convert DWARF type generation to use loader
DO NOT SUBMIT

Begin converting the the portion of DWARF generation that emits types
to use the new loader interfaces. Creation of subprogram DIES and
compilation unit DIEs still operate on sym.Symbols at the moment.
This code enabled under the linker flag "-newdw".

This version of the code will make it through dwarfGenerateDebugInfo(),
e.g. DWARF type generation. What is left to be done is a walk through
all fo the DIE objects at the point of loader.LoadFull to convert
any embedded loader.Sym refs in DIEs to sym.Symbols.

Change-Id: I681a00fb8a1f3c37884a79b373d86411332e07c8