[dev.link] cmd/link: convert DWARF type generation to use loader
DO NOT SUBMIT

Begin converting the the portion of DWARF generation that emits types
to use the new loader interfaces. Creation of subprogram DIES and
compilation unit DIEs still operate on sym.Symbols at the moment.
This code enabled under the linker flag "-newdw".

This is a work in progress, still a fair amount of implementation is
missing.

Change-Id: I681a00fb8a1f3c37884a79b373d86411332e07c8