runtime: do not use PowerRegisterSuspendResumeNotification on systems with "program time" timer
Systems where PowerRegisterSuspendResumeNotification returns 2 are also
systems where nanotime() is on "program time" rather than "real time",
since this error code is for places where the kernel can't get any
information about suspend and resume. Thus, in this case, we don't want
the notifier anyway, so we make it non-fatal here. This applies, for
example, to Docker users.

Fixes #35447
Fixes #35482

Change-Id: I9e1ce5bbc54b9da55ff7a3918b5da28112647eee