cmd/go: fix and re-enable build_trimpath test
The test was comparing a binary built from a list of files to a test
build from a named package. That should not (and did not) work. The
test now compares two binaries built the same way in different
directories.

Also, fix the gccgo portion of the test. It should build from the same
path within two different GOPATH roots, not from two paths within the
same root.

Fixes #35435

Change-Id: I2535a0011c9d97d2274e5550ae277302dbb91e6f