go/parser: rewrite with cmd/compile/internal/syntax
A proof-of-concept that passes some of the tests, and kind of works.

cmd/compile/internal/syntax isn't importable, so copy it to
internal/syntax. Keep the original package as-is, to reduce the size of
the diff. The copy is completely unmodified, as the exported API is good
enough for go/parser.

Not all go/parser APIs are supported yet, such as ParseExpr*.

Change-Id: I6e3676aca6620060e24a5d38af2d980220930c27