Update patch set 1
Patch Set 1:

I wrote the bulk of this in September 2018, after a brief exchange with Robert where we wondered if this would be possible to do in go/parser directly without losing features or performance.

I ended up stashing the project, but I've rebased it and cleaned it up a bit over the past few days.

Ian, I seem to remember you mentioning that gopls had its own tweaks on go/parser to report better errors and accept more kinds of incomplete source. I haven't been able to locate that right now, so any pointers are appreciated.

In the meantime, I'll leave this here and continue trying to make the tests happier and finish up the API.

Patch-set: 1
CC: Gerrit User 9711 <9711@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 22687 <22687@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 5260 <5260@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 7392 <7392@62eb7196-b449-3ce5-99f1-c037f21e1705>