Update patch set 2
Patch Set 2:

Cherry and James Knight mentioned this to me earlier this evening, I took the trouble to fire off a quick 5-builds 5-runs benchmark to compare version 2 of the patch to its parent.  Results here:

https://perf.golang.org/search?q=upload:20191127.1

The notable changes were in
  build user time (geomean +3.54%, many slowdowns)
  overall binary size (geomean +1.47%)

The reported benchmark geomean was +0.13% (time), but there were mixed winners and losers.
This was on hardware that has no microcode changes (yet).

Patch-set: 2
CC: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>