Update patch set 2
Patch Set 2:

ENFILE is the title of issue #35131. This patch aims at closing it as only the Windows part is missing. 

This is how I found that generation didn't work as expected and the relevant issue #32309.
I suggest to tackle it separately as the WSA* messages are currently unavailable in zerrors_windows.go. The code for the error messages for windows seems unavailable in the x/sys repo.

Patch-set: 2