cmd/go: make cgo pkg-config work w/ lib paths that have spaces
In 2016, an issue was raised that #cgo pkg-config: annotations did not
work with paths that have spaces in them.  The relevant issue is #16455.
This was fixed, but only for CFLAGS.  Because a Windows user was not
available, the windows use case was overlooked (linux is unlikely to
have libs folders with spaces).

The same logic developed for #16455 is now applied to --libs output in
addition to --cflags output, and tests have been created for the
getPkgConfigFlags method in exec.go.

Fixes #35262

Change-Id: Id4440a8415259254fde759b720e6239ecd3138e0
GitHub-Last-Rev: 82f75fc8a62ba389fe794ef5ff6b89cef4dd3aea
GitHub-Pull-Request: golang/go#35909