Update patch set 4
Patch Set 4:

(2 comments)

Almost 7000 error messages are automatically generated in x/sys/windows. Only a handful are relevant for Go and are located in syscall/types_windows. There does not seem to be any rationale to any of them. Assignment of values currently in x/sys would not add anything. 

Submitted change has no impact on function .Is(errno) as the error code returned remains the Windows error code. For instance:
The directory is not empty. != syscall.ENOTEMPTY which is "directory not empty"
145 != syscall.ENOTEMPTY which return code 536871002
There is maybe some confusion on the purpose as the mapping of some error codes are summarized by the .Is() method. Error.As as referred in issue #29054 might be the answer.

The proposed commit:
- Map additionnal Windows Errno to corresponding dummy Go codes. Text must be removed.
- Add WSAENOBUFS and WSAEMFILE error codes to mark them as temporary which should close issue #35131 (and #15637, its predecessor).
- Add windows build tag which seems missing in touched _windows files.

Error codes are added in syscall because Temporary() method is defined there. Moving error messages to internal/syscall/windows remained elusive. It creates some interlocking which fail the build.

It might be useful to consider the correction reported in h2_bundle for WSACONNECT* error codes.

Patch-set: 4