cmd/compile: change uint64/int64/float64 in structs to be 8-byte aligned on 32-bit architectures
Prototype of a change to address issue #599. The main code changes are in
cmd/compile/internal/gc/align.go and cmd/compile/internal/types/type.go, but
adjustments in a few other places as well.

The main issue encountered (besides Cgo compatibility) is that two Linux Syscall
structures on 32-bit architectures, Stat_t and Flock_t (e.g., see
ztypes_linux_386.go), have a 64-bit field which is not aligned to 64-bits
naturally. Therefore, changing the alignment rules according to this issue would
change the layout of those structs. For testing purposes, I changed the relevant
fields to be two 32-bit fields, which allows the syscalls to work, but breaks the
ABI.

But, other than the ABI change because of those two structs, all tests work on
linux-amd64, linux-386, linux-arm64, and linux-arm.

To demonstrate the change is working, I removed padding fields from a bunch of
runtime structs and from WaitGroup.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I5bf0537799041dca2a2cf39d2bf3093611d9fe88