[dev.link] cmd/link: prepopulate oReader rcache
During the loader's Preload and LoadRefs methods, the loader is
already making a sweep through the non-package def and non-package ref
sections of each Go object's local index space, determining which
global symbol to which the local syms map. Given that we're doing
lookups for all of these locals and determining the Sym's to which
they map, it makes sense to prepopulate the oReader's local -> global
cache with the results. This brings the rcache hit rate up to
effectively 100% during dead code elimination.

Change-Id: If2e872e113f200bffc7c7f72c645355da812a1d6