[dev.link] cmd/link: convert ELF host object loading to new loader
This is a rewrite of the ELF host object loader to use just the Loader
interfaces for symbol creation, without constructing sym.Symbols. At
the moment this is gated under the temporary linker command line
option "-newldelf".

This version is able to get through make.bash on linux/amd64 but not
yet all of all.bash (has a couple of problems with misc/cgo).

Change-Id: I99f41368f75b0df9e35ef3c2cf2a702b732540c6