runtime: revert "monitor for suspend/resume to kick timeouts"
This reverts CL 191957 as well as the follow-up CL 198417, "runtime:
iterate ms via allm linked list to avoid race".

There are two different clocks we're concerned with:

"Real time", aka CLOCK_BOOTTIME on Linux or "interrupt time" on
Windows. This measures the passage of real time and continues to pass
when the system is suspended. This clock has meaning in the external
world, such as to users and across networks and distributed systems.

"Program time", aka CLOCK_MONOTONIC on POSIX or "unbiased interrupt
time" on Windows. This also measures the passage of real time, but
pauses when the system is suspended and no programs can make progress.
This clock is more meaningful internal to a system.

Neither of these are affected by wall clock adjustments. "Wall time"
is another clock, but not one we're concerned with here.

Prior to Windows 8, sleeps were in terms of real time. Windows 8
changed the system behavior so sleeps were in terms of program time.
Sleeps on POSIX systems have always been in terms of program time. CL
191957 attempted to change the Windows behavior to always be in terms
of real time by monitoring system suspends.

However, this change had some issues. It did not converge the behavior
between Windows and Unix; if anything it made it more divergent. The
change broke Go applications running Docker because the suspend
monitoring fails (#35447), and the only obvious "fix" for that problem
seems to be to ignore some errors, which would lead to sleep behavior
that is mostly real time, but sometimes program time. Finally, the
change also ignores some sleep states, again leading to inconsistent
sleep behavior.

Given all this, this CL reverts the change and returns us to the state
where on Windows >= 8, sleeping is in terms of program time. The old
state is not the "right" answer in any absolute sense, but we’d been
in the old state for years with few problems and the new state broke
more users than it helped and is even more unpredictably inconsistent
both between OSes and between environments and setups on Windows. I
think the whole approach to clocks bears further consideration, but
1.14 testing has shown that we don't have the right answer yet.

Updates #31528.
Fixes #35447.
Fixes #35482.

Change-Id: I215c6d8aee16e7f742fadb2d2f5b44102a35abdd