Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> As explained in the GH issue for this, this will introduce a regression that breaks network protocol software, such as WireGuard. Rather, there's work to be done instead on using CLOCK_BOOTTIME on other operating systems.

I am aware. I explained my reasoning for this change on the GitHub issue, which was a consensus of several people on the Go team after quite a bit of review (including looking over the WireGuard source). There is no consensus that all of the OSes should move to CLOCK_BOOTTIME semantics, as evidenced by golang.org/issue/24595, and the day before the intended 1.14 beta is not the time to make that decision or to attempt a complex implementation change.

I recognize this is not ideal for WireGuard.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>