Update patch set 1
Patch Set 1:

> Patch Set 1:
> I am aware. I explained my reasoning for this change on the GitHub issue,

Sorry, I hadn't seen that last night. I just replied to it today.

> which was a consensus of several people on the Go team after quite a bit of review (including looking over the WireGuard source)

Of those people on the Go team, how many of them have devoted entire weeks to improving Windows support on Go? If Alex was a part of that decision, for example, I'd be inclined to respect it. Otherwise, this doesn't bode well.

> There is no consensus that all of the OSes should move to CLOCK_BOOTTIME semantics, as evidenced by golang.org/issue/24595

Maybe yes, maybe no, but you don't have a choice with Windows 7, which we support.

> I recognize this is not ideal for WireGuard.

It basically signals to us, "move to a different runtime/language/platform," since the protocol simply _does-not-work_ otherwise, and intentionally breaking it doesn't project much confidence. On the other hand, I could be wrong, and maybe you have a great way of making wireguard-go work with CLOCK_MONOTONIC semantics? Or with half-BOOTTIME/half-MONOTONIC semantics, which is what this revert commit will induce. If so, please do send a patch to us.

Patch-set: 1
Reviewer: Gerrit User 20556 <20556@62eb7196-b449-3ce5-99f1-c037f21e1705>